Get rate and fraction data

Get rate and fraction data

Usage

get_rates_and_fractions(
  x,
  sites = NULL,
  catchments = NULL,
  condition = NULL,
  icd10_regex = NULL,
  maternal = FALSE,
  cond_name_short = condition[1],
  causal_chain = TRUE,
  adjust_vars_override = NULL,
  factor_groups = NULL,
  pval_cutoff = 0.1,
  pct_na_cutoff = 20,
  per = 10000
)

Arguments

x

Processed CHAMPS dataset.

sites

A vector of site names to include in the calculations. If NULL, all sites with data will be used.

catchments

A vector of catchments to include in the calculations. If NULL, all catchments with data will be used.

condition

CHAMPS group specifying the condition

icd10_regex

An optional regular expression specifying ICD10 codes that define a condition.

maternal

Are we searching for a maternal condition or ICD10 specification? Default is FALSE.

cond_name_short

The name of the condition to use in outputs (e.g. if the condition is "Congenital birth defects", cond_name_short could be "CBD"). Defaults to condition if not specified.

causal_chain

if TRUE, the causal chain is searched, if FALSE, the underlying cause is searched

adjust_vars_override

An optional vector of adjustment variables that will override the automatic adjustment. Cannot be more than two variables and must include age.

factor_groups

A named list that specifies how to group factors

pval_cutoff

P-value for which variables are considered for adjustment. Ignored if adjust_vars_override is specified.

pct_na_cutoff

Proportion of missingness for which variables are considered for adjustment. Ignored if adjust_vars_override is specified.

per

Calculate rates and fractions per how many births. 10k is the default.

Note

One or both of icd10_regex and condition must be specified